cobalt/registry: return -EAGAIN upon lack of registry slot
-ENOMEM is confusing in this case, does not actually reflect the error
condition, and does not match the error code commonly returned by
POSIX services for denoting a (temporary) lack of resources.
Besides, this source of error was not even mentioned in the
documentation of the affected services.

All error codes must be detected, and any program that might have
specifically checked for -ENOMEM during error recovery in the affected
services was potentially confused, so this change does not introduce a
significant ABI variation.